+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Added `iconParams` for customize icons used within the grid.
 - Modified to support `display` as callback method.
 - Reimplemented `headerSpan` from v1.
+- Modified the data type of `uniqueIndex` of event callback argument to number format.
 
 ## 2.0.0 (Apr 24, 2019)
 - Removed dependency on jQuery and jQuery UI and compatible with modern web frameworks such as Bootstrap 4.